rag_pipeline_experimentation/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-pipeline-experimentation"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "rag-common",
    "pymupdf>=1.23",
    "llm-utils",
    "openai>=1.0",
    "instructor>=1.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pandas>=2.0",
    "numpy>=1.24",
    "tiktoken>=0.5",
    "streamlit>=1.25",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov"]

[tool.uv.sources]
rag-common = { path = "../rag_common", editable = true }
llm-utils  = { path = "../llm_utils",  editable = true }

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
